Scene2 commit 0
Program runs as expected. Box is transported from A to B. Start button is functional and requires a 3sec hold before the conveyor begins. While it is held, a siren is sounded. Basic annunciator are implemented for each state of the tool. I chose to use a word in memory D0 to hold the tool state and each state time is tracked to ensure the expected timeline. Specific error handling code is not implemented.